Rename thread directory and move READMEs.
